quickstart/requirements.txt
python-dotenv
graphrag_sdk[openai]
beautifulsoup4
requests
